// src/champion/types/champion.type.ts
export type ChampionTypes = {
  version: string;
  id: string;
  key: string;
  name: string;
  title: string;
  blurb: string;
  info: ChampionInfoTypes;
  image: ChampionImageTypes;
  tags: string[];
  partype: string;
  stats: ChampionStatsTypes;
};

export type ChampionInfoTypes = {
  attack: number;
  defense: number;
  magic: number;
  difficulty: number;
};
export type ChampionImageTypes = {
  full: string;
  sprite: string;
  group: string;
  x: number;
  y: number;
  w: number;
  h: number;
};

export type ChampionStatsTypes = {
  hp: number;
  hpperlevel: number;
  mp: number;
  mpperlevel: number;
  movespeed: number;
  armor: number;
  armorperlevel: number;
  spellblock: number;
  spellblockperlevel: number;
  attackrange: number;
  hpregen: number;
  hpregenperlevel: number;
  mpregen: number;
  mpregenperlevel: number;
  crit: number;
  critperlevel: number;
  attackdamage: number;
  attackdamageperlevel: number;
  attackspeedperlevel: number;
  attackspeed: number;
};